Add polyline_loops_have_boundaries option.
Added an option to control whether a polyline loop is considered to
have a non-empty boundary:

  S2BooleanOperation::Option::polyline_loops_have_boundaries()

By default this option is true, meaning that even when the first and
last vertices of a polyline are the same, the polyline is still
considered to have a well-defined "start" and "end".  For example, if
the polyline boundary model is OPEN then the polyline loop would not
include the start/end vertices.  These are the best semantics for most
applications, such as GPS tracks or road network segments.

If the polyline forms a loop and this option is set to false, then
instead the first and last vertices are considered to represent a
single vertex in the interior of the polyline.  In this case the
boundary of the polyline is empty, meaning that the first/last vertex
will be contained by the polyline even if the boundary model is OPEN.