Add S2CellIndex.
S2CellIndex stores a set of (cell_id, label) pairs.  It can be used to
index a collection of S2CellUnions, and then find the S2CellUnions that
are closest to a given target object (using S2ClosestCellQuery, which is
forthcoming).

The index has built-in methods for finding the S2CellUnions that
intersect a given target S2CellUnion.  This is a very fast operation,
especially when the output size is small.  For example, for an index
with 10,000 S2Cap coverings, finding the coverings that intersect a
given target covering takes about 200ns plus 200ns per intersected
covering (using coverings with 16 cells).