Add S2ClosestCellQuery.
Add a new class S2ClosestCellQuery that efficiently finds the closest
cell(s) to a given target within a given indexed collection of cells.
The indexed cells are represented as an S2CellIndex, while the target
geometry can be a point, edge, S2Cell, S2CellUnion, or S2ShapeIndex (an
arbitrary collection of points, polylines, and polygons).

A typical use case would be to add a collection of S2Cell coverings to
an S2CellIndex, and then use S2ClosestCellQuery to find all coverings
that are within a given distance of some target geometry.

Here is an example showing how to find all cells that are closer than
"distance_limit" to a given target point:

  S2ClosestCellQuery query(&cell_index);
  query.mutable_options()->set_max_distance(distance_limit);
  S2ClosestCellQuery::PointTarget target(target_point);
  for (const auto& result : query.FindClosestCells(&target)) {
    // result.distance() is the distance to the target.
    // result.cell_id() is the indexed S2CellId.
    // result.label() is a label associated with the S2CellId.
    DoSomething(target_point, result);
  }

This commit also adds a new target type for distance queries
(CellUnionTarget) that can be used to measure the distance to an
S2CellUnion.  The new target type can be used with S2ClosestCellQuery,
S2ClosestPointQuery, or S2ClosestEdgeQuery.