Fix S2EdgeTessellator antimeridian wrapping.
Fix S2EdgeTessellator so that it wraps geometry that passes through the
antimeridian correctly.  Previously it did the correct wrapping
calculations for individual edges but not for sequences of edges.  Now
for example, if you tessellate an edge chain that wraps around the
equator multiple times, then the longitude coordinates will increase or
decrease monotonically.