S2Builder: Treat polygons with no edges as empty.
Change S2Builder so that by default, polygons with no edges are assumed
to be empty.  Previously the user was expected to supply a predicate in
this situation to determine whether the polygon is empty or full.
Eventually there will be standard predicates that can be used in most
situations, but for now it is unrealistic to expect users to do this.

The main advantage of relaxing this requirement is that it allows
s2builderutil::S2PolygonLayer and s2builderutil::LaxPolygonLayer to
construct both empty and full polygons by calling the supplied predicate
(since now the default predicate simply says that the polygon is empty
rather than returning an error).

Also add quite a few more comments about why these predicates even exist.