Optimize EncodedUintVector<T>::lower_bound
Templatize the binary search loop to have a separate case for each
encoded integer length (1 - 8 bytes).  This produces optimal code for
accessing integers in the inner loop, e.g. a 4-byte integer will be
accessed using a single instruction where as 5-byte integer will be
accessed using two loads, a shift, and a bitwise or.

This makes lower_bound 20-60% faster.