Add operator new/delete and ctor for Candidate (#124)
Inspired by #85, which reports a gcc ubsan violation in AddCandidate.

I was not able to reproduce with -fsanitize=undefined but the current code
does look suspicious to me.

Current code:
Candidate* candidate = static_cast<Candidate*>(
    ::operator new(sizeof(Candidate) + children_size));

However, the Candidate constructor is never called, and Candidate is not
trivially constructible.  Candidate contains S2Cell, which contains
R1Interval, which is not trivially constructible.

Add Candidate::operator new with an extra arg of max_children;
this delegates to ::operator new.  Similarly add Candidate::operator delete.

Move the Candidate initialization logic to a constructor.