SWIG: Expose s2builderutil::S2PolygonLayer and S2::S2BufferOperation
The S2BufferOperationOptions class is a workaround for being unable to
wrap nested classes with SWIG. The SWIG documentation suggests it can
be done, but I couldn't get it to work, and comments in s2_common.i
suggest that others haven't been able to either.

I had similar difficulty with S2Shape, which I worked around by hiding
it from the Python interface entirely.