Fix macOS Python build error (#399)
Mac seems to use dynamic linking more for making modules than Linux and
Windows do. Thus the common python symbols were missing and it requires
libpython3.13.dylib for these symbols. If this is a different in the mac
"Framework" then cmake should probably include that for the
Development.Module component. But this is a workaround in the meantime
while we wait for any upstream fixes.

Closes #398

Signed-off-by: Andy Neff <andy@visionsystemsinc.com>