test.yml: build with --parallel 45 (#429)
This was disabled for Linux because the jobs were getting cancelled.  It
is far too slow building sequentially (8 minutes instead of 4 minutes).
It's a bit disappointing that we only have a 2x speedup.
Limit parallel jobs to 45 since more seems trigger resource limit termination.

--parallel alone means unlimited parallelism.

Also update build.yml for consistency, but that already had --parallel.