Add bignum.h and port exactfloat to use it.
This removes the dependency on OpenSSL for its bignum library. The main benefit
of this is that targetting things like WASM are much simpler since we don't need
to compile OpenSSL to WASM.

I didn't try as hard for performance optimization as OpenSSL (which also has a
hard requirement around constant-time operation for security purposes), but
benchmarks indicate addition and subtraction is generally the same and
multiplication is within a factor of two. For the size of bignums we expect in
S2 there shouldn't be a noticeable difference in performance.