S2Cell::GetDistance: Optimize same-face case
Use the uv coordinates to prune the vertex-edge distance computations
needed.

When cell A is above B, we don't need to compare the top vertices/edge
of A with the bottom vertices/edge of B.

Note, however, that when A is both above and left of B, we cannot just
compute the lower-right/upper-left vertex distance due to the
projection.

It is possible that more comparisons can be omitted; I'm not sure.

Error is within 1e-15 radians of previous results.

GetDistance shows a 2x speedup for same-face cells.