S2Cell::GetDistance: Optimize same-face case (#479)
3.1x speedup for same-face cells.

Use the uv coordinates to prune the vertex-edge distance computations needed. Only 4 vertex-edge distance computations need to be made instead of 32, and 2 of those can be faster interior distance computations.

We can reduce the number of comparisons since, when cell A is above B, we don't need to compare the top vertices/edge of A with the bottom vertices/edge of B. When A is both above and left of B, we just need to compare the direction with the greater uv distance.

The approach is based on ericv's explanation in https://github.com/google/s2geometry/pull/479#issuecomment-3586683570

Error is within 1e-15 radians of previous results.